+++ components/src/dynamo/common/multimodal/__init__.py
@@ -11,11 +11,10 @@ from dynamo.common.multimodal.embedding_transfer import (
     NixlPersistentEmbeddingSender,
     TransferRequest,
 )
-from dynamo.common.multimodal.image_batch_loader import load_image_batch
 from dynamo.common.multimodal.image_loader import ImageLoader
 
 __all__ = [
-    "AsyncEncoderCache", "ImageLoader", "load_image_batch",
+    "AsyncEncoderCache", "ImageLoader",
     "NixlPersistentEmbeddingReceiver",
     "NixlPersistentEmbeddingSender",
     "TransferRequest",
