+++ components/src/dynamo/common/multimodal/image_loader.py
@@ -18,14 +18,22 @@ import base64
 import binascii
 import logging
 from io import BytesIO
+from typing import Any, Dict, Final, List, Optional
 from urllib.parse import urlparse
 
 import httpx
 from PIL import Image
 
+import dynamo.nixl_connect as nixl_connect
+from dynamo.common.utils.media_nixl import read_decoded_media_via_nixl
+
 from .http_client import get_http_client
 
 logger = logging.getLogger(__name__)
+
+# Constants for multimodal data variants
+URL_VARIANT_KEY: Final = "Url"
+DECODED_VARIANT_KEY: Final = "Decoded"
 
 
 class ImageLoader:
@@ -108,3 +116,77 @@ class ImageLoader:
         except Exception as e:
             logger.error(f"Error loading image: {e}")
             raise ValueError(f"Failed to load image: {e}")
+
+    async def load_image_batch(
+        self,
+        image_mm_items: List[Dict[str, Any]],
+        enable_frontend_decoding: bool = False,
+        nixl_connector: Optional["nixl_connect.Connector"] = None,
+    ) -> List[Any]:
+        """
+        Load a batch of images from multimodal data items.
+
+        Supports two paths:
+        1. Url variant: Download and decode image from URL (default)
+        2. Decoded variant: Read pre-decoded image via NIXL RDMA (requires enable_frontend_decoding=True)
+
+        Args:
+            image_mm_items: List of multimodal data items for images
+            enable_frontend_decoding: If True, enables NIXL RDMA for decoded images
+            nixl_connector: NIXL connector for frontend decoding (required if enable_frontend_decoding=True)
+
+        Returns:
+            List of loaded image data
+
+        Raises:
+            Exception: If any image fails to load
+            ValueError: If enable_frontend_decoding=True but nixl_connector is None
+        """
+        image_futures = []
+
+        for item in image_mm_items:
+            if isinstance(item, dict) and URL_VARIANT_KEY in item:
+                # URL path: download and decode in Python backend
+                url = item[URL_VARIANT_KEY]
+                image_futures.append(self.load_image(url))
+                logger.debug(f"Preparing to load image from URL: {url[:80]}...")
+            elif isinstance(item, dict) and DECODED_VARIANT_KEY in item:
+                if enable_frontend_decoding:
+                    if nixl_connector is None:
+                        logger.error(
+                            "Frontend decoding enabled but nixl_connector not provided. "
+                            "Caller must pass an initialized NIXL connector."
+                        )
+                        raise ValueError(
+                            "nixl_connector required when enable_frontend_decoding=True"
+                        )
+
+                    metadata = item[DECODED_VARIANT_KEY]
+                    image_futures.append(
+                        read_decoded_media_via_nixl(nixl_connector, metadata)
+                    )
+                else:
+                    logger.error(
+                        "Received Decoded multimodal data but enable_frontend_decoding=False. "
+                        "Set enable_frontend_decoding=True to enable NIXL RDMA image transfer."
+                    )
+                    raise ValueError("Could not load decoded media from frontend")
+
+        # Process images in parallel
+        results = await asyncio.gather(*image_futures, return_exceptions=True)
+        loaded_images = []
+        collective_exceptions = ""
+        for media_item, result in zip(image_mm_items, results):
+            if isinstance(result, Exception):
+                source = media_item.get(URL_VARIANT_KEY, "decoded")
+                logger.error(f"Failed to load image from {source[:80]}...: {result}")
+                collective_exceptions += (
+                    f"Failed to load image from {source[:80]}...: {result}\n"
+                )
+                continue
+            loaded_images.append(result)
+
+        if collective_exceptions:
+            raise Exception(collective_exceptions)
+
+        return loaded_images
